refactor(dashboards): join reload and cadence into one split button
Reload and the cadence dropdown were two separate outline buttons sitting
next to each other, which is what made the cadence half ambiguous — an
adjacent control is just another control, and beside the time-range
picker it read as a second time control.

Grafana attaches them, and attachment is what carries the meaning: "30s"
welded to "Reload" can only be read one way. So the two halves are now
one split button sharing a seam, and the cadence half is back to showing
just the interval (or "Off") rather than explaining itself with a word.

`RefreshControls` takes the reload action as a prop and `PageRefreshControls`
binds it to the page refresh context, because the share page drives its
own refresh and has no such context. `ReloadControls` stays as-is for
traces, metrics, service-map and infra, which have no cadence to offer.
